+++ database/factories/Visitors/VisitorFactory.php
@@ -26,6 +26,7 @@ class VisitorFactory extends Factory
         return [
             'name' => $this->faker->name($gender),
             'id_number' => $this->faker->optional(0.6)->numberBetween(10000, 99999),
+            'membership_number' => $this->faker->optional(0.6)->numberBetween(10000, 99999),
             'gender' => $gender,
             'date_of_birth' => $this->faker->dateTimeBetween('-50 years', '-1 month'),
             'nationality' => $this->faker->optional(0.95)->country,